feat(markdown-panel): add logic to save and edit buttons
